+++ pakyow-core/lib/pakyow/generators/project/example.rb
@@ -1,9 +1,5 @@
 # frozen_string_literal: true
 
 generator :project, :example, extends: :project do
-  action :generate_default_project do |destination_path, **options|
-    Generators::Project.new(
-      ::File.expand_path("../../../generatable/project/default", __FILE__)
-    ).generate(destination_path, **options)
-  end
+  source_path File.expand_path("../../../generatable/project/example", __FILE__)
 end